Change the sequence of request every 3s information.

+++ Services/HushClient.ApplicationSettings/Model/BlockchainInfo.cs
@@ -2,5 +2,5 @@ namespace HushClient.ApplicationSettings.Model;
 
 public class BlockchainInfo
 {
-    public int LastHeightSynched { get; set; }    
+    public double LastHeightSynched { get; set; }    
 }
